Update BUILD.md
This is just to ask the question as to the specific drive we want it installed on.  Do we install it on disk0 - The internal Hard Drive or do we install it on Disk 2 for example (only other one I see) that has the image we just created.  IF this page could be clarified a little bit more for us not so savvy people.  Thank you